Paradigm Support Calculates to what extent the cognitive map signals support for a set of two commensurable policy paradigms. More specifically, it determines the saliency of all concepts classified as aligning with one of the two paradigms which are valued positively as determined via the evaluate_concepts function. As incommensurability suggests that there is a zero-sum relation between the paradigms, negative valued concepts belonging to paradigm a will be interpreted as support for paradigm b. As such the function lists the saliency value of negatively valued concepts as a positive score for the rival paradigm. To be able to run this function, the functions calculate-degrees AND evaluate_concepts should have been run first. Also the nodelist (node_measures) list this function takes needs to contain a column with the name 'paradigms' in which for each concept it is noted with which of the two paradigms it aligns. If a concept does speak to either of the rival paradigms, the cell should be left empty. The researcher can use any rivaling set of paradigms that is relevant to their research as long as no more or less than two are included.

Paradigm Support Calculates to what extent the cognitive map signals support for a set of two commensurable policy paradigms. More specifically, it determines the saliency of all concepts classified as aligning with one of the two paradigms which are valued positively as determined via the evaluate_concepts function. As incommensurability suggests that there is a zero-sum relation between the paradigms, negative valued concepts belonging to paradigm a will be interpreted as support for paradigm b. As such the function lists the saliency value of negatively valued concepts as a positive score for the rival paradigm. To be able to run this function, the functions calculate-degrees AND evaluate_concepts should have been run first. Also the nodelist (node_measures) list this function takes needs to contain a column with the name 'paradigms' in which for each concept it is noted with which of the two paradigms it aligns. If a concept does speak to either of the rival paradigms, the cell should be left empty. The researcher can use any rivaling set of paradigms that is relevant to their research as long as no more or less than two are included.

Usage

paradigm_support(node_measures, paradigm_a, paradigm_b)

Arguments

node_measures

An object of class "dataframe" including all measures returned when running the functions calculate_degrees & evaluate_concepts, as well as a column with the title "paradigms" in which the concepts are classified as belonging to a set of two pre-determined paradigms. If a concept does not refer to either paradigm, the cell should be left empty.

paradigm_a

An object of class "character string" / the first name of a paradigm that occurs in the node_measures dataframe.

paradigm_b

An object of class "character string" / the first name of a paradigm that occurs in the node_measures dataframe. To derive the exact names of the paradigms as they appear in the node_measures dataframe, run the following code: "paradigms <- unique(node_measures$paradigms) #deriving the names of the two paradigms from the node_measures dataframe paradigms <- na.omit(paradigms) #omitting the empty cells (NULL category) from the analysis"

Value

Returns an object of class "dataframe" with additional columns with the paradigms as column titles and the saliency scores for those concepts that indicate a positive stance towards the paradigm.

Examples

"paradigms <- base::unique(rutte_p2_node_measures$paradigms) #deriving all instrument-types
from the node_measures dataframe
paradigms <- base::na.omit(paradigms) #omitting the empty cells (NULL category)
from the analysis"
#> [1] "paradigms <- base::unique(rutte_p2_node_measures$paradigms) #deriving all instrument-types\nfrom the node_measures dataframe\nparadigms <- base::na.omit(paradigms) #omitting the empty cells (NULL category)\nfrom the analysis"
rutte_p2_node_measures <- paradigm_support(rutte_p2_node_measures, paradigm_a, paradigm_b)
#> Error in dplyr::case_when(node_measures$paradigms == paradigm_a & node_measures$val_run1 >     0 ~ node_measures$w_degree, node_measures$paradigms == paradigm_b &     node_measures$val_run1 < 0 ~ node_measures$w_degree): Failed to evaluate the left-hand side of formula 1.
#> Caused by error:
#> ! object 'rutte_p2_node_measures' not found