rutte_p2_node_measures

This is an intermediate dataset used for examples in the vignette

Usage

rutte_p2_node_measures

Format

A data frame with 84 rows and 14 variables:

id

num numerical id for all unique concepts in the associated edgelist

node_name

chr name of all unique concepts in the associated edgelist

paradigms

chr the categorisation of the concept as belonging to/being associated with one of two incommensurable paradigms

int

chr the categorisation of the concept as intergovernmental or supranational (superfluous column for now)

value

num the intrinsic (normative) value of the concept displayed as a number: 1 = positive/neutral, -1 = negative.

instruments

chr the categorisation of the concept as a particular type of policy-instrument. This column can also be used for any other type of categorization of the concepts that is deemed relevant

indegree

indegree

outdegree

outdegree

degree

degree

w_indegree

w_indegree

w_outdegree

w_outdegree

w_degree

w_degree

go

go

gow

gow

Source

van Esch, F., & Snellens, J. (2024). How to ‘measure’ ideas. Introducing the method of cognitive mapping to the domain of ideational policy studies. Journal of European Public Policy, 31(2), 428–451. https://doi.org/10.1080/13501763.2022.2155215